separate conditional average from ofio
