+++ conditionalMean/compute_conditionalMean.py
@@ -9,6 +9,7 @@ import pickle
 
 from bcr_util import *
 from ofio import *
+from mathtools import *
 
 parser = argparse.ArgumentParser(
     description="Compute conditional means of OpenFOAM fields"
